Makefile: Optimize for sm8150's Kryo 485 CPU and use -O3 on Clang
Clang doen't support optimizing for two clusters at once, so help the
weaker little cluster out a bit by optimizing exclusive for it instead
since it needs all the speed it can get.

Optimizing with -O3 is unsafe with GCC and sometimes causes odd issues,
but is known to be safe with Clang. Enable it when compiling with Clang
for some extra micro-optimizations.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>